requirements.txt
opencv-python
pytesseract
mss
numpy
Pillow
